requirements.txt
pandas
pyodbc
Office365-REST-Python-Client
pymongo 
pyodbc 
office365-rest-python-client